refactor: 발급 기한 검증 책임을 coupon으로 이동
