+++ src/test/java/coupon/membercoupon/domain/MemberCouponTest.java
@@ -3,14 +3,12 @@ package coupon.membercoupon.domain;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import coupon.CouponException;
 import coupon.coupon.domain.Coupon;
 import coupon.fixture.CouponFixture;
 import coupon.fixture.MemberFixture;
 import coupon.member.domain.Member;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 class MemberCouponTest {
 
@@ -26,20 +24,5 @@ class MemberCouponTest {
 
         //then
         assertThat(memberCoupon.getIssuedAt().plusDays(7)).isBeforeOrEqualTo(memberCoupon.getExpiredAt());
-    }
-
-    @DisplayName("쿠폰 발급 기한 외의 날짜에 발급을 시도하면 예외가 발생한다.")
-    @Test
-    void cannotIssue() {
-        // given
-        LocalDate startAt = LocalDate.now().plusDays(1);
-        LocalDate endAt = LocalDate.now().plusDays(10);
-        Coupon coupon = CouponFixture.create(startAt, endAt);
-        Member member = MemberFixture.create();
-
-        // when
-        assertThatThrownBy(() -> MemberCoupon.issue(member, coupon))
-                .isInstanceOf(CouponException.class)
-                .hasMessage("쿠폰을 발급할 수 있는 기한이 지났습니다.");
     }
 }
